[X] Create Book directory for BookCards Components. [X] Move to utils.ts the getPorcentajeBook for BookCard animation.
